requirements.txt
# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
Flask>=3.0.0
Flask-Cors>=4.0.0

# Database
aiosqlite>=0.19.0
sqlalchemy>=2.0.0
lancedb>=0.3.0

# AI/ML
google-generativeai>=0.8.0
plotly>=5.20.0
matplotlib>=3.8.0
seaborn>=0.13.0
PyMySQL>=1.1.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=12.0.0
scikit-learn>=1.3.0
cachetools>=5.3.0
PyYAML>=6.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
aiofiles>=23.0.0

# HTTP
aiohttp>=3.8.0

# Optional (for document processing)
PyMuPDF>=1.23.0
Pillow>=10.0.0
geopandas>=0.14.0
Shapely>=2.0.0
